Move quaternion functions to quaternion namespace
